Добавлена старница с товарами Добавлены страницы с добавлением, редактирование и удалением товаров.

+++ by.pvt.api/src/main/java/by/pvt/api/dto/ProductResponse.java
@@ -1,17 +1,17 @@
 package by.pvt.api.dto;
 
 public class ProductResponse {
-    private int id;
+    private long id;
     private String type;
     private String nameProduct;
     private int codeProduct;
     private double price;
 
-    public ProductResponse(Integer ID, String type, String nameProduct, Integer codeProduct, double PRice){
+    public ProductResponse(long ID, String type, String nameProduct, Integer codeProduct, double PRice){
 this.id = ID; this.type = type; this.nameProduct = nameProduct; this.codeProduct = codeProduct; this.price = PRice;
 }
 public ProductResponse(){}
-public int getId(){ return id; }
+public long getId(){ return id; }
 public String getType(){
 return type;
 }
@@ -24,7 +24,7 @@ return codeProduct;
 public double getPrice(){
 return price;
 }
-    public void setId(Integer Id){ this.id = Id; }
+    public void setId(long Id){ this.id = Id; }
     public void setType(String t){ this.type = t;}
     public void setNameProduct(String nP){ this.nameProduct = nP; }
     public void setCodeProduct(Integer cp){ this.codeProduct = cp;  }
